+++ src/test/common.ts
@@ -1,0 +1,23 @@
+export class MockEventEmitter<T> {
+  private listeners: ((e: T) => any)[] = [];
+
+  event = (listener: (e: T) => any) => {
+    this.listeners.push(listener);
+    return {
+      dispose: () => {
+        const index = this.listeners.indexOf(listener);
+        if (index > -1) {
+          this.listeners.splice(index, 1);
+        }
+      },
+    };
+  };
+
+  fire(data: T): void {
+    this.listeners.forEach((listener) => listener(data));
+  }
+
+  dispose(): void {
+    this.listeners = [];
+  }
+}
