+++ webpack.config.js
@@ -45,7 +45,7 @@ const config = {
     rules: [
       {
         test: /\.ts$/,
-        exclude: /node_modules/,
+        exclude: /(node_modules|src\/test)/,
         use: [
           {
             loader: "ts-loader",
